Migrate DB calls to the new CasinoDB class
